refactor home to login component
