+++ src/app/public/home/home.component.html
@@ -1,71 +1,1 @@
-<form [formGroup]="loginForm" (ngSubmit)="onSubmit(loginForm.value)" fxFlex fxLayout="column"
-  fxLayoutAlign="center center">
-
-  <div fxFlex></div>
-
-  <div style="min-width: 300px;" fxLayout="column" fxLayoutGap="40px">
-
-    <p class="productive-heading-05">Log in to xlear list</p>
-
-    <div fxLayout="column" fxLayoutGap="20px" *ngIf="!showPasswordBlock; else passwordBlock">
-
-      <div class="bx--form-item" fxLayout="column">
-        <label for="email" class="bx--label">
-          Email
-        </label>
-        <input formControlName="email" id="email" type="text" class="bx--text-input" placeholder="you@company.com">
-      </div>
-
-      <fieldset class="bx--fieldset">
-        <ibm-checkbox (change)="onRememberMe($event)">
-          Remember Me
-        </ibm-checkbox>
-      </fieldset>
-
-      <div class="bx--form-item">
-        <button class="bx--btn bx--btn--primary" type="button"
-          (click)="showPasswordBlock = !showPasswordBlock">Continue</button>
-      </div>
-
-    </div>
-
-    <ng-template #passwordBlock>
-
-      <div fxLayout="column" fxLayoutGap="20px">
-
-        <div class="bx--form-item">
-          <label for="password" class="bx--label">Password</label>
-          <input formControlName="password" id="password" type="password" class="bx--text-input" placeholder="Password">
-        </div>
-
-        <div class="bx--form-item">
-          <a href="#" ibmLink>forgot password?</a>
-        </div>
-
-        <div class="bx--form-item" *ngIf="!loginAttemptInProgress; else loginAttemptBlock">
-          <button class="bx--btn bx--btn--primary" type="button" type="submit">Log in</button>
-        </div>
-
-        <ng-template #loginAttemptBlock>
-          <ibm-loading size="sm"></ibm-loading>
-        </ng-template>
-
-      </div>
-
-    </ng-template>
-
-    <div *ngIf="loginErrorNotification">
-      <ibm-notification [notificationObj]="loginErrorNotification">
-      </ibm-notification>
-    </div>
-
-    <div fxLayout="column">
-      <span class="helper-text-01">Don't have an account?</span>
-      <a [routerLink]="[ '/register']" ibmLink>Create an account</a>
-    </div>
-
-  </div>
-
-  <div fxFlex></div>
-
-</form>
+home works!
